==> code/02-atac/17a_histone_chip.ipynb <==
library(rtracklayer)
library(tidyverse)
library(ChrAccR)
library(motifmatchr)
library(SummarizedExperiment)
library(ggpubr)
library(patchwork)

### read all TF data: cpm, motif matching

deseq_dir <- "../../output/02-atac/04/deseq_data_Plate_TF_Dose/"
plotdir <- "../../plots/02-atac/17/"
outdir <- "../../output/02-atac/17/"

dir.create(plotdir, showWarnings = F, recursive=T)
dir.create(outdir, showWarnings = F, recursive=T)

# read in ATAC CPM values
cpm_df <- read.csv("../../output/02-atac/01/cpm.tsv", sep=" ") # just use the cpm normalized peak count matrix
cpm_df %>% head
dim(cpm_df)

# read the motif matching results
motif.scores <- readRDS("../../output/02-atac/03/motif.scores.JASPAR2020.rds") # this is JASPAR motifs
motif.scores

if (is.null(rownames(motif.scores))){
    # read in the motif peak names
    dsa_norm <- loadDsAcc('../../output/02-atac/01/dsa_norm_rpkmlog2quantile/')
    region <- "filteredConsensus"

    Peaks <- as.data.frame(dsa_norm@coord[[region]],row.names = NULL, stringsAsFactors=FALSE) %>%
    dplyr::select(c("seqnames","start","end","name"))
    rownames(motif.scores) <- Peaks$name
    Peaks <- GRanges(Peaks)
}

### inputs

# inputs
motif_code_ls <- c(
                   #"ALX4"="MA0681.2_PHOX2B",
                   #"ELF1"="MA0473.3_ELF1",
                   #"IRF4"="MA1419.1_IRF4",
                   "KLF1"="MA0039.4_KLF4",
                   "KLF4"="MA0039.4_KLF4",
                   #"LEF1"="MA0768.1_LEF1",
                   #"NR4A1"="MA1112.2_NR4A1",
                   #"OCT4"="MA1115.1_POU5F1",
                   #"PRDM1"="MA0508.3_PRDM1",
                   "SOX2"="MA0143.4_SOX2",
                   #"SP4"="MA0685.1_SP4",
                   "SPI1"="MA0080.5_SPI1")
                   #"TCF3"="MA0522.3_TCF3")

group1 = "GFP_1" # reference level
region = "filteredConsensus"

i = 3
TF = names(motif_code_ls)[i]
group2 = paste0(TF, "_1")
motif_code = motif_code_ls[i]

### read TF specific ATAC CPM, motif matching, differential analysis and sensitivity analysis results

cpm <- cpm_df[,c(grep("GFP", colnames(cpm_df)), grep(TF, colnames(cpm_df)))]
cpm %>% head
dim(cpm)

# identify peaks containing target motif
target_motif_counts <- rowSums(motifCounts(motif.scores[,motif_code]))
subset <- motif.scores[(target_motif_counts > 0), motif_code]
meta <- read.table(paste0("../../output/02-atac/10/", TF,"_motif_containing_peak_meta.tsv"), row.names=1, sep="\t")
rowData(subset) <- meta[rownames(subset),]
subset

### compare median ATAC CPM for highest dose sample in different peak groups

maxdose_cols <- colnames(cpm)[grep("100", colnames(cpm))]
idx <- which(rowData(subset)$saturating=="Yes")
means1 <- cpm[rownames(subset)[idx],maxdose_cols] %>% rowMeans
f1 <- means1 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at highest plasmid dose \nsaturating sensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means1),2)))

idx <- which(rowData(subset)$saturating=="No")
means2 <- cpm[rownames(subset)[idx],maxdose_cols] %>% rowMeans
f2 <- means2 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at highest plasmid dose \nnonsaturating sensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means2),2)))

idx <- which(rowData(subset)$state=="open")
means3 <- cpm[rownames(subset)[idx],maxdose_cols] %>% rowMeans
f3 <- means3 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at highest plasmid dose \nopen nonsensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means3),2)))

idx <- which(rowData(subset)$state=="closed")
means4 <- cpm[rownames(subset)[idx],maxdose_cols] %>% rowMeans
f4 <- means4 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at highest plasmid dose \nclosed nonsensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means4),2)))

df <- c(means1,means2,means3,means4) %>% as.data.frame
colnames(df) <- "ATAC_CPM"
df$group <- c(rep("saturating sensitive", length(means1)),
             rep("nonsaturating sensitive", length(means2)),
             rep("open nonsensitive", length(means3)),
             rep("closed nonsensitive", length(means4)))
f5 <- ggboxplot(df, x="group", y="ATAC_CPM") + ggtitle(paste0(TF, " ATAC CPM at highest TF dose")) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
f6 <- ggboxplot(df, x="group", y="ATAC_CPM") + ggtitle(paste0(TF, " ATAC CPM at highest TF dose")) + coord_cartesian(ylim = c(-1, 10)) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))


pdf(paste0(plotdir, "/",TF,"_",motif_code,"_ATAC_CPM_highdose.pdf"), width=5, height=5)
lapply(paste0("f",1:6),function(n){get(n)})
dev.off()

### compare median ATAC CPM for control sample in different peak groups

mindose_cols <- colnames(cpm)[grep("GFP", colnames(cpm))]
idx <- which(rowData(subset)$saturating=="Yes")
means1 <- cpm[rownames(subset)[idx],mindose_cols] %>% rowMeans
f1 <- means1 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at 0 plasmid dose \nsaturating sensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means1),2)))

idx <- which(rowData(subset)$saturating=="No")
means2 <- cpm[rownames(subset)[idx],mindose_cols] %>% rowMeans
f2 <- means2 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at 0 plasmid dose \nnonsaturating sensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means2),2)))

idx <- which(rowData(subset)$state=="open")
means3 <- cpm[rownames(subset)[idx],mindose_cols] %>% rowMeans
f3 <- means3 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at 0 plasmid dose \nopen nonsensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means3),2)))

idx <- which(rowData(subset)$state=="closed")
means4 <- cpm[rownames(subset)[idx],mindose_cols] %>% rowMeans
f4 <- means4 %>% gghistogram + 
    ggtitle(sprintf("distribution of ATAC CPM at 0 plasmid dose \nclosed nonsensitive peaks \n(n=%s, median=%s)", length(idx), round(median(means4),2)))

df <- c(means1,means2,means3,means4) %>% as.data.frame
colnames(df) <- "ATAC_CPM"
df$group <- c(rep("saturating sensitive", length(means1)),
             rep("nonsaturating sensitive", length(means2)),
             rep("open nonsensitive", length(means3)),
             rep("closed nonsensitive", length(means4)))
f5 <- ggboxplot(df, x="group", y="ATAC_CPM") + ggtitle(paste0(TF, " ATAC CPM at 0 TF dose")) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
f6 <- ggboxplot(df, x="group", y="ATAC_CPM") + ggtitle(paste0(TF, " ATAC CPM at 0 TF dose")) + coord_cartesian(ylim = c(-1, 10)) + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

pdf(paste0(plotdir, "/", TF,"_",motif_code,"_ATAC_CPM_nodose.pdf"), width=5, height=5)
lapply(paste0("f",1:6),function(n){get(n)})
dev.off()

### ChIP overlap with peak groups
download from ENCODE, list of ChIP-seq datasets are in 17c

plot_overlap_chip <- function(bed, overlap, name){
    nhit <- length(overlap)
    bedsub <- bed[overlap@to]
    medsig <- bedsub$signalValue %>% median %>% round(2)
    medq <- bedsub$qValue %>% median %>% round(2)

    pp1 <- bedsub$signalValue %>% gghistogram + 
        ggtitle(paste0(mod, " ChIP signalValue distribution\n", name, " peaks \nnoverlap=", nhit," median=", medsig)) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    pp2 <- bedsub$qValue %>% gghistogram + 
        ggtitle(paste0(mod, " ChIP -log10(qValue) distribution\n", name, " peaks\nnoverlap=", nhit," median=", medq)) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    return(list(pp1,pp2))
}

# single block for easy running
sample <- read.csv("17b_encode_sample_info.csv")

for (mod in c("H3K27ac", "H3K9me3", "H3K4me1", "H3K4me3", "H3K36me3-1", "H3K36me3-2")){
    bed <- read.table(paste0(outdir, "/",sample[sample$mod==mod, "bedfile"]), sep="\t")
    colnames(bed) <- c("chrom", "chromStart", "chromEnd", "name","score","strand","signalValue","pValue","qValue","peak")
    bed <- GRanges(bed)
    # macs2 narrow peaks output: signal value - measurement of overall enrichment for the region, 
    #                            pValue - statistical significance (-log10), qValue - Statistical significance using FDR (-log10), 
    #                            peak - Point-source called for this peak; 0-based offset from chromSTart
    
    plots <- list()
    
    # overlap each group of peaks with ChIP signal
    query1 <- subset@rowRanges[which(rowData(subset)$saturating=="Yes")]
    overlap1 <- findOverlaps(query1, bed)
    plots <- c(plots, plot_overlap_chip(bed, overlap1, "saturating sensitive"))
    
    query2 <- subset@rowRanges[which(rowData(subset)$saturating=="No")]
    overlap2 <- findOverlaps(query2, bed)
    plots <- c(plots, plot_overlap_chip(bed, overlap2, "nonsaturating sensitive"))
    
    query3 <- subset@rowRanges[which(rowData(subset)$state=="open")]
    overlap3 <- findOverlaps(query3, bed)
    plots <- c(plots, plot_overlap_chip(bed, overlap3, "open nonsensitive"))
    
    query4 <- subset@rowRanges[which(rowData(subset)$state=="closed")]
    overlap4 <- findOverlaps(query4, bed)
    plots <- c(plots, plot_overlap_chip(bed, overlap4, "closed nonsensitive"))
    
    df <- c(bed[overlap1@to]$signalValue,bed[overlap2@to]$signalValue,bed[overlap3@to]$signalValue,bed[overlap4@to]$signalValue) %>% data.frame
    colnames(df) <- "signalValue"
    df$qValue <- c(bed[overlap1@to]$qValue,bed[overlap2@to]$qValue,bed[overlap3@to]$qValue, bed[overlap4@to]$qValue)
    df$group <- c(rep("saturating_sensitive", length(bed[overlap1@to]$signalValue)),
                 rep("nonsaturating_sensitive", length(bed[overlap2@to]$signalValue)),
                 rep("open_nonsensitive", length(bed[overlap3@to]$signalValue)),
                 rep("closed_nonsensitive", length(bed[overlap4@to]$signalValue)))
    df$peakname <- c(names(query1[overlap1@from]), names(query2[overlap2@from]), names(query3[overlap3@from]), names(query4[overlap4@from]))

    p9 <- ggplot(df, aes(x = group)) + geom_bar(stat = "count") + theme_classic() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + ylab("# hits")
    
    
    # plot percentage of peaks with the histone mod in each group
    groups <- c("saturating_sensitive", "nonsaturating sensitive", "open_nonsensitive", "closed_nonsensitive")
    pct <- data.frame(group=factor(groups, levels=groups),
                     total_peaks=c(length(query1),length(query2), length(query3), length(query4)),
                      mod_peaks=c(length(unique(overlap1@from)),length(unique(overlap2@from)),
                                 length(unique(overlap3@from)),length(unique(overlap4@from))))
    pct$pct_mod_peaks <- round(pct$mod_peaks / pct$total_peaks * 100, 2)
    p10 <- ggbarplot(pct, x="group",y="pct_mod_peaks",label = T, color="darkgray",fill = "darkgray") + ylab(paste0("% of peaks with ",mod)) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    
    # plot number of peaks per group
    p11 <- ggbarplot(pct, x="group",y="total_peaks",label = T, color="darkgray",fill = "darkgray") + ylab("# peaks in group") + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    
    # cross group comparison
    p12 <- ggboxplot(df,x="group",y="signalValue") + ylab(paste0(mod, " ChIP signalValue")) + xlab(paste0(motif_code, " motif containing peaks in ", TF, " samples")) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    p13 <- ggboxplot(df,x="group",y="signalValue") + ylab(paste0(mod, " ChIP signalValue")) + xlab(paste0(motif_code, " motif containing peaks in ", TF, " samples")) + coord_cartesian(ylim = c(-1, 30)) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    
    p14 <- ggboxplot(df,x="group",y="qValue") + ylab(paste0(mod, " ChIP -log10(qValue)")) + xlab(paste0(motif_code, " motif containing peaks in ", TF, " samples")) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    p15 <- ggboxplot(df,x="group",y="qValue") + ylab(paste0(mod, " ChIP -log10(qValue)")) + xlab(paste0(motif_code, " motif containing peaks in ", TF, " samples")) + coord_cartesian(ylim = c(-1, 50)) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

    # just comparing saturating vs nonsaturating sensitive peaks
    dff <- df[df$group %in% c("saturating_sensitive","nonsaturating_sensitive"),]
    p16 <- ggboxplot(dff,x="group",y="signalValue") + ylab(paste0(mod, " ChIP signalValue")) + xlab(paste0(motif_code, " motif containing peaks in ", TF, " samples")) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    p17 <- ggboxplot(dff,x="group",y="signalValue") + ylab(paste0(mod, " ChIP signalValue")) + xlab(paste0(motif_code, " motif containing peaks in ", TF, " samples")) + coord_cartesian(ylim = c(-1, 20)) + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
    
    # correlating EC50 and hill coef of saturating sensitive peaks with chip data
    dff <- df[df$group=="saturating_sensitive",]
    dff$EC50 <- rowData(subset)[dff$peakname, "EC50"]
    dff$h <- rowData(subset)[dff$peakname, "h"]
    
    p18 <- ggplot(data=dff, aes(y=log10(EC50),x=log10(signalValue)))+ geom_point() + theme(text = element_text(size=20)) +
      geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))
    p19 <- ggplot(data=dff, aes(y=log10(h),x=log10(signalValue)))+ geom_point() + theme(text = element_text(size=20)) +
          geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))
    p20 <- ggplot(data=dff, aes(y=log10(EC50),x=log10(qValue)))+ geom_point() + theme(text = element_text(size=20)) +
          geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))
    p21 <- ggplot(data=dff, aes(y=log10(h),x=log10(qValue)))+ geom_point() + theme(text = element_text(size=20)) +
          geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))

    p22 <- ggplot(data=dff, aes(y=EC50,x=signalValue))+ geom_point() + theme(text = element_text(size=20)) +
      geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))
    p23 <- ggplot(data=dff, aes(y=h,x=signalValue))+ geom_point() + theme(text = element_text(size=20)) +
          geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))
    p24 <- ggplot(data=dff, aes(y=EC50,x=qValue))+ geom_point() + theme(text = element_text(size=20)) +
          geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))
    p25 <- ggplot(data=dff, aes(y=h,x=qValue))+ geom_point() + theme(text = element_text(size=20)) +
          geom_smooth(method='lm', formula= y~x) + stat_cor(aes(label =  ..rr.label..))


    pdf(paste0(plotdir, "/", TF,"_",motif_code,"_",mod,".pdf"), width=5, height=5)
    for (p in plots){print(p)}
    lapply(paste0("p",9:25),function(n){print(get(n))})
    dev.off()
}
